+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Support the following commands of the API v3.0:
    - `VLCCommandNameNext`: Jumps to next playlist item.
    - `VLCCommandNamePlay`: Plays a specific playlist item.
+   - `VLCCommandNamePrevious`: Jumps to previous playlist item.
    - `VLCCommandNameSeek`: Seek to a particular point in the current playback
    - `VLCCommandNameStatus`: Retrieve the player status informations
    - `VLCCommandNameStop`: Stops the player to play the current media